src/requirements.txt
## The following requirements were added by pip --freeze:
Flask==0.10.1
Flask-Admin==1.0.8
Flask-Bootstrap==3.1.1.2
Flask-GoogleLogin==0.3.1
Flask-KVSession==0.4
Flask-Login==0.2.10
Flask-Moment==0.3.1
Flask-SQLAlchemy==1.0
Flask-Sijax==0.3.3
Flask-SocketIO==0.3.4
Flask-WTF==0.9.5
Flask-WhooshAlchemy==0.55
Jinja2==2.7.2
MarkupSafe==0.21
PyICL==0.6.15
SQLAlchemy==0.9.4
Sijax==0.2.7
WTForms==1.0.5
Werkzeug==0.9.4
Whoosh==2.6.0
argparse==1.1
beautifulsoup4==4.3.2
blinker==1.3
boto==2.27.0
cookbook==2.2.25
dateutils==0.6.6
gevent==1.0.1
gevent-socketio==0.3.6
gevent-websocket==0.9.3
google-api-python-client==1.2
greenlet==0.4.2
httplib2==0.9
itsdangerous==0.24
npyscreen==3.37
pyRFC3339==0.2
python-dateutil==2.2
pytz==2014.2
requests==0.14.2
simplejson==3.4.1
simplekv==0.8.4
six==1.6.1
virtualenv==1.11.5
wsgiref==0.1.2
